+++ Assets/Scripts/FlowStates/FlowMessages/FlowGroup.cs
@@ -6,10 +6,10 @@ namespace FlowStates.FlowMessages
     public class FlowGroup : MonoBehaviour
     {
         private FlowStateMachine m_fsm;
-        private byte m_flowStateId;
+        private int m_flowStateId;
         private byte m_flowWindowId;
 
-        public void Link(FlowStateMachine fsm, byte flowStateId, byte flowWindowId = byte.MaxValue)
+        public void Link(FlowStateMachine fsm, int flowStateId, byte flowWindowId = byte.MaxValue)
         {
             m_fsm = fsm;
             m_flowStateId = flowStateId;
